+++ src/content-script/components/QueryBox/QueryBox.tsx
@@ -1,15 +1,12 @@
 import CommonSelectors from "@/content-script/components/QueryBox/CommonSelectors";
 import ImageModelSelector from "@/content-script/components/QueryBox/ImageModelSelector";
 import useFetchUserSettings from "@/content-script/hooks/useFetchUserSettings";
-import { useInit } from "@/content-script/hooks/useInit";
 import useInitQueryBoxSessionStore from "@/content-script/hooks/useInitQueryBoxSessionStore";
 import useQueryBoxObserver from "@/content-script/hooks/useQueryBoxObserver";
 import useCplxGeneralSettings from "@/cplx-user-settings/hooks/useCplxGeneralSettings";
 import Portal from "@/shared/components/Portal";
 
 export default function QueryBox() {
-  const { isWebSocketCaptured } = useInit();
-
   const { settings } = useCplxGeneralSettings();
 
   const { focus, imageGenModel, languageModel, collection } =
@@ -61,8 +58,7 @@ export default function QueryBox() {
   });
 
   const isReady =
-    (isWebSocketCaptured && !!userSettings && !isLoadingUserSettings) ||
-    !!userSettingsFetchError;
+    (!!userSettings && !isLoadingUserSettings) || !!userSettingsFetchError;
 
   const selectors = (
     <CommonSelectors