+++ src/types/ws.types.ts
@@ -1,3 +1,8 @@
+import {
+  MessageData,
+  WebSocketEventData,
+} from "@/types/webpage-messenger.types";
+
 export type WsParsedMessage = {
   messageCode: number;
   event: string;
@@ -11,4 +16,15 @@ export type RouterEvent =
 
 export function isParsedWsMessage(data: unknown): data is WsParsedMessage {
   return data != null && typeof data === "object" && "messageCode" in data;
+}
+
+export function isWebSocketEventData(
+  data: unknown,
+): data is MessageData<WebSocketEventData> {
+  return (
+    data != null &&
+    typeof data === "object" &&
+    "event" in data &&
+    data.event === "webSocketEvent"
+  );
 }